requirements.txt
annotated-types==0.7.0
anyio==4.5.2
astunparse==1.6.3
black==24.8.0
click==8.1.8
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.115.6
graphql-core==3.2.5
h11==0.14.0
idna==3.10
motor==3.6.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.4
pydantic-core==2.27.2
pymongo==4.9.2
python-dateutil==2.9.0.post0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
strawberry-graphql==0.255.0
tomli==2.2.1
typing-extensions==4.12.2
uvicorn==0.33.0
